CU-1uvzyp - Updated endpoints and some code cleanup

+++ src/Astro.API/Application/Request/Post/CelestialPostRequestModel.cs
@@ -24,8 +24,8 @@ namespace Astro.API.Application.Request.Post
 
         public string Description { get; set; }
 
-        public decimal Distance { get; set; }
+        public double Distance { get; set; }
 
-        public decimal DistanceTolerance { get; set; }
+        public double DistanceTolerance { get; set; }
     }
 }
